requirements.txt
streamlit
psutil
plotly
mysql-connector-python
streamlit_autorefresh 
python-dotenv